+++ Changelog.md
@@ -1,1 +1,3 @@
-+ Added additional parameter to requests which bypasses BoxSets created by the TVDB BoxSet plugin.
+* Improved Series support (+ Season objects)
+* Clarified help message
+* Reset your database, if you don't want to get spammed. Last time I didn't consider the usefulness of Season objects :/